+++ loader.php
@@ -2789,6 +2789,7 @@ if ( !function_exists( 'radio_station_load_prefixed_functions' ) ) {
 // =========
 
 // == 1.1.4 ==
+// - fix for debug prefix key dash
 // - fix for weird get_settings glitch bug (isset failing?!)
 
 // == 1.1.3 ==