+++ CHANGELOG.md
@@ -7,6 +7,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### 2.3.3.5
+* Fixed: day left/right shifting on Schedule table/tab mobile views
+* Added: past/today/future filter for Schedule Override List 
+* Added: filter for Schedule display start day (and to accept today)
+
 ### 2.3.3.4
 * Improved: auto-match show description to info height on Show pages
 * Improved: allow multiple Related Show selection for single post